Started change to output file
Had the idea to output all the statistics to a .csv that way the data
can be easily imported into Excel and would not need to be written to
tons of different files that all need a specialized parser. It
shouldn’t be a big change to the code. I will work on this tomorrow
hopefully.